Refactor run loop and agent utilities
